+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   (this is useful when importing a backup of an account you already have and can't distinguish which is the old account and which is the imported one)
 - Add context menu to gallery
 - Option for packagers to disable asar (`NO_ASAR=true npm run pack:generate_config`).
+- Added context menu for info messages
 
 ### Changed
 - Change "More info" translation to "Message Details"